requirements.txt
gradio>=3.50.0
networkx>=3.1
matplotlib>=3.7.0
pygraphviz>=1.10
pillow>=9.0.0
numpy>=1.24.0